Reorganize the interoperability configurations: move the import of InteroperabilityConfig and LoggerConfig, remove commonConfig, and update the exports in the main configuration file.
