+++ packages/commons/src/config/index.ts
@@ -1,7 +1,9 @@
-export * from "./commonConfig.js";
 export * from "./jwtConfig.js";
 export * from "./signerConfig.js";
 export * from "./databaseConfig.js";
 export * from "./shConfig.js";
 export * from "./shClientConfig.js";
 export * from "./m2mConfig.js";
+export * from "./interoperabilityConfig.js";
+export * from "./httpServerConfig.js";
+export * from "./loggerConfig.js";
